Get rid of basically all build warnings with the new system of typing used throughout.

+++ Code/net/net.h
@@ -116,7 +116,7 @@ namespace hemelb
 
         void CreateMPIType(ProcComms *iMetaData);
 
-        void EnsureEnoughRequests(unsigned int count);
+        void EnsureEnoughRequests(size_t count);
 
         bool sendReceivePrepped;
